+++ UmbracoVault/IUmbracoContext.cs
@@ -124,6 +124,15 @@ namespace UmbracoVault
         void FillClassProperties<T>(T instance, Func<string, bool, object> getPropertyValue);
 
         /// <summary>
+        /// Fills out class properties based on a provided, instantiated class and a Func instructing it how to get the raw property data based on alias.
+        /// This allows the TypeHandler system to be used by related classes.
+        /// </summary>
+        /// <typeparam name="T">Type of instantiated class</typeparam>
+        /// <param name="instance">An newed-up instance of T</param>
+        /// <param name="getPropertyValue">A func that, provided a property alias, a Property Info, and a boolean to indicate recursion, will return a raw value to be processed by the TypeHandler system</param>
+        void FillClassProperties<T>(T instance, Func<string, PropertyInfo, bool, object> getPropertyValue);
+
+        /// <summary>
         /// Uses Vault context to try to resolve value for a single property
         /// </summary>
         /// <param name="getPropertyValue">Func to return raw value for the property</param>
@@ -131,6 +140,15 @@ namespace UmbracoVault
         /// <param name="value">If found value will be set</param>
         /// <returns>True if able to resolve value, false if not</returns>
         bool TryGetValueForProperty(Func<string, bool, object> getPropertyValue, PropertyInfo propertyInfo, out object value);
+
+        /// <summary>
+        /// Uses Vault context to try to resolve value for a single property
+        /// </summary>
+        /// <param name="getPropertyValue">Func to return raw value for the property</param>
+        /// <param name="propertyInfo">Property for which to return a value</param>
+        /// <param name="value">If found value will be set</param>
+        /// <returns>True if able to resolve value, false if not</returns>
+        bool TryGetValueForProperty(Func<string, PropertyInfo, bool, object> getPropertyValue, PropertyInfo propertyInfo, out object value);
 
         /// <summary>
         /// Given a class, will return true if the class is intended to be hydrated as a Media object instead of a Content object