+++ CHANGELOG.md
@@ -1,4 +1,35 @@
+## v2.0.0 (UNRELEASED)
+
+### Behavioral Change
+
+#### Child-Type casting
+
+Asking Vault to provide you an item by a specific type will now test to see if the item has a child type it can be converted to. If it does, it constructs it as the child item type and returns the item cast as the parent item type. For example,
+
+```csharp
+public class Foo { }
+
+public class Bar : Foo { }
+
+...
+// If the item can be a Bar object, 
+// Vault will construct it as a `Bar` object and hydrate `Bar` properties. 
+
+Vault.GetItem<Foo>();
+```
+
+## v1.3.0 (UNRELEASED)
+
+**New Feature: Independence from Web Context**
+
+Extended the current UmbracoContext implementation to include a secondary implementation that removes any dependency on `IPublishedContext`, `UmbracoHelper`, and the Umbraco WebContext.  Instead of using the WebContext as a vehicle for retrieving data from Umbraco, the [ServiceContext](https://our.umbraco.org/documentation/Reference/Management/Services/) is used. [#27]
+
+**Other**
+ 
+ * Added `PropertyInfo` parameter to `FillClassProperties` to support additional behavior for libraries that extend Vault.  [#25]
+
 ## v1.2.1
+
  * Fixes an issue where classes that used type inheritance would cause duplicate results to return in calls to `GetUmbracoEntityAliasesFromType` [#12]
 
 ## v1.2.0